+++ src/main/java/com/buuz135/industrial/block/resourceproduction/tile/OreLaserBaseTile.java
@@ -223,10 +223,10 @@ public class OreLaserBaseTile extends IndustrialMachineTile<OreLaserBaseTile> im
 
     @Override
     public boolean canAcceptAugment(ItemStack augment) {
-        if (AugmentWrapper.hasType(augment, AugmentTypes.SPEED)) {
-            return false;
+        if (AugmentWrapper.hasType(augment, AugmentTypes.EFFICIENCY)) {
+            return true;
         }
-        return super.canAcceptAugment(augment);
+        return false;
     }
 
     @Override