+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# Version 3.6.11
+
+* Added Pink Slime Block
+* Fixed addon filter in the fluid/ore laser base
+* Fixed fluid/ore laser base progress bar going over the limit in big setups
+
 # Version 3.6.10
 
 * Fixed manual recipe